Extract DPM PM operating quality contracts
